refactor: made sizes of shape double instead of int

+++ src/main/kotlin/io/github/agentseek/common/Circle2d.kt
@@ -4,7 +4,7 @@ package io.github.agentseek.common
  * This class represents a Circle in two dimensions, with [position] equal to its top left corner
  * and [radius]
  */
-data class Circle2d(val radius: Int, override var position: Point2d = Point2d(0.0, 0.0)) : Shape2d {
+data class Circle2d(val radius: Double, override var position: Point2d = Point2d(0.0, 0.0)) : Shape2d {
     override var center: Point2d
         get() = position
         set(value) {
